[debugger] add Autobreak support
Autobreak is a feature I have named for when you want to look at the data for a breakpoint which needs to occur potentially thousands of times, or more than you would want to do manually.

I had experimented with updating the GUI but it was happening so slow that I finalized on a terminal-based approach. This means you must start bsnes from a terminal. Only tested on OSX, it can be done like this: bsnes.app/Contents/MacOS/bsnes

Once you set your breakpoint, and activate the autobreak checkbox, the emulator will run in realtime as your breakpoint data gets logged in the terminal.. FYI, use Cmd-K to clear the console window before logging data.