[ui-qt / snes: breakpoint] Add "numbreaks" feature
Numbreaks allows you to specify an interval of breakpoints to occur before actually stopping code execution and reporting information to the debugger. For example, if you set numbreaks to 4 on a write to $0 in PPU OAM, the debugger would only halt after 4 writes have been made to the OAM. In the case of OAM breakpoints, no further information is provided in the debugger window. But, if this had been a CPU breakpoint, details will be printed for only the 4th break - the one that caused the halt.

It is possible to further code bsnes to log the "non-breaking" breakpoints to the terminal. The terminal is preferred over the GUI, as it is such with "auto-break," for maintaining speed. (Frequently breaking via GUI is slow!!). However, bsnes could be further configured to use either the terminal, GUI, or both. [feature-request]

To satisfy the previous 2 feature requests stated in the prior paragraph, it may be satisfying to add a "Debugger Options" sub-menu in bsnes Options menu.